added callbacks

+++ src/quick/java/eu/siacs/conversations/ui/VerifyActivity.java
@@ -1,5 +1,6 @@
 package eu.siacs.conversations.ui;
 
+import android.app.AlertDialog;
 import android.content.ClipData;
 import android.content.ClipDescription;
 import android.content.ClipboardManager;
@@ -12,30 +13,31 @@ import android.support.v7.widget.Toolbar;
 import android.text.Html;
 import android.view.View;
 
-import java.util.ArrayList;
-
 import eu.siacs.conversations.R;
 import eu.siacs.conversations.databinding.ActivityVerifyBinding;
 import eu.siacs.conversations.entities.Account;
+import eu.siacs.conversations.services.QuickConversationsService;
 import eu.siacs.conversations.ui.util.PinEntryWrapper;
 import eu.siacs.conversations.utils.AccountUtils;
 import eu.siacs.conversations.utils.PhoneNumberUtilWrapper;
 
 import static android.content.ClipDescription.MIMETYPE_TEXT_PLAIN;
 
-public class VerifyActivity extends XmppActivity implements ClipboardManager.OnPrimaryClipChangedListener {
+public class VerifyActivity extends XmppActivity implements ClipboardManager.OnPrimaryClipChangedListener, QuickConversationsService.OnVerification {
 
     private ActivityVerifyBinding binding;
     private Account account;
     private PinEntryWrapper pinEntryWrapper;
     private ClipboardManager clipboardManager;
     private String pasted = null;
+    private boolean verifying = false;
 
 
     @Override
     protected void onCreate(final Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         String pin = savedInstanceState != null ? savedInstanceState.getString("pin") : null;
+        boolean verifying = savedInstanceState != null && savedInstanceState.getBoolean("verifying");
         this.pasted = savedInstanceState != null ? savedInstanceState.getString("pasted") : null;
         this.binding = DataBindingUtil.setContentView(this, R.layout.activity_verify);
         setSupportActionBar((Toolbar) this.binding.toolbar);
@@ -44,16 +46,57 @@ public class VerifyActivity extends XmppActivity implements ClipboardManager.OnP
             this.pinEntryWrapper.setPin(pin);
         }
         binding.back.setOnClickListener(this::onBackButton);
+        binding.next.setOnClickListener(this::onNextButton);
         clipboardManager = (ClipboardManager) getSystemService(Context.CLIPBOARD_SERVICE);
+        setVerifyingState(verifying);
     }
 
     private void onBackButton(View view) {
+        if (this.verifying) {
+            setVerifyingState(false);
+            return;
+        }
+        final Intent intent = new Intent(this, EnterPhoneNumberActivity.class);
         if (this.account != null) {
-            xmppConnectionService.deleteAccount(account);
-            Intent intent = new Intent(this, EnterPhoneNumberActivity.class);
+            AlertDialog.Builder builder = new AlertDialog.Builder(this);
+            builder.setMessage(R.string.abort_registration_procedure);
+            builder.setPositiveButton(R.string.yes, (dialog, which) -> {
+                xmppConnectionService.deleteAccount(account);
+                startActivity(intent);
+                finish();
+            });
+            builder.setNegativeButton(R.string.no, null);
+            builder.create().show();
+        } else {
             startActivity(intent);
             finish();
         }
+    }
+
+    private void onNextButton(View view) {
+        final String pin = pinEntryWrapper.getPin();
+        if (PinEntryWrapper.isValidPin(pin)) {
+            if (account != null && xmppConnectionService != null) {
+                setVerifyingState(true);
+                xmppConnectionService.getQuickConversationsService().verify(account, pin);
+            }
+        } else {
+            AlertDialog.Builder builder = new AlertDialog.Builder(this);
+            builder.setMessage(R.string.please_enter_pin);
+            builder.setPositiveButton(R.string.ok, null);
+            builder.create().show();
+        }
+    }
+
+    private void setVerifyingState(boolean verifying) {
+        this.verifying = verifying;
+        this.binding.back.setText(verifying ? R.string.cancel : R.string.back);
+        this.binding.next.setEnabled(!verifying);
+        this.binding.next.setText(verifying ? R.string.verifying : R.string.next);
+        this.binding.resendSms.setVisibility(verifying ? View.GONE : View.VISIBLE);
+        pinEntryWrapper.setEnabled(!verifying);
+        this.binding.progressBar.setVisibility(verifying ? View.VISIBLE : View.GONE);
+        this.binding.progressBar.setIndeterminate(verifying);
     }
 
     @Override
@@ -63,16 +106,19 @@ public class VerifyActivity extends XmppActivity implements ClipboardManager.OnP
 
     @Override
     void onBackendConnected() {
+        xmppConnectionService.getQuickConversationsService().addOnVerificationListener(this);
         this.account = AccountUtils.getFirst(xmppConnectionService);
         if (this.account == null) {
             return;
         }
         this.binding.weHaveSent.setText(Html.fromHtml(getString(R.string.we_have_sent_you_an_sms, PhoneNumberUtilWrapper.prettyPhoneNumber(this, this.account.getJid()))));
+        setVerifyingState(xmppConnectionService.getQuickConversationsService().isVerifying());
     }
 
     @Override
     public void onSaveInstanceState(Bundle savedInstanceState) {
         savedInstanceState.putString("pin", this.pinEntryWrapper.getPin());
+        savedInstanceState.putBoolean("verifying", this.verifying);
         if (this.pasted != null) {
             savedInstanceState.putString("pasted", this.pasted);
         }
@@ -89,6 +135,9 @@ public class VerifyActivity extends XmppActivity implements ClipboardManager.OnP
     public void onStop() {
         super.onStop();
         clipboardManager.removePrimaryClipChangedListener(this);
+        if (xmppConnectionService != null) {
+            xmppConnectionService.getQuickConversationsService().removeOnVerificationListener(this);
+        }
     }
 
     @Override
@@ -105,7 +154,7 @@ public class VerifyActivity extends XmppActivity implements ClipboardManager.OnP
             final ClipData primaryClip = clipboardManager.getPrimaryClip();
             if (primaryClip != null && primaryClip.getItemCount() > 0) {
                 final CharSequence clip = primaryClip.getItemAt(0).getText();
-                if (PinEntryWrapper.isPin(clip) && !clip.toString().equals(this.pasted)) {
+                if (PinEntryWrapper.isValidPin(clip) && !clip.toString().equals(this.pasted)) {
                     this.pasted = clip.toString();
                     pinEntryWrapper.setPin(clip.toString());
                     final Snackbar snackbar = Snackbar.make(binding.coordinator, R.string.possible_pin, Snackbar.LENGTH_LONG);
@@ -122,5 +171,17 @@ public class VerifyActivity extends XmppActivity implements ClipboardManager.OnP
         if (pinEntryWrapper.isEmpty()) {
             pastePinFromClipboard();
         }
+    }
+
+    @Override
+    public void onVerificationFailed() {
+        runOnUiThread(() -> {
+            setVerifyingState(false);
+        });
+    }
+
+    @Override
+    public void onVerificationSucceeded() {
+
     }
 }
